requirements.txt
pyyaml>=6.0
openai>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
